Split utils in store
